# apps/api/src/openapi.yaml
openapi: 3.0.3
info:
  title: Arkeo Racing API
  description: API for the Arkeo Racing game - a horse racing simulation powered by blockchain service provider metrics.
  version: 0.1.0
  contact:
    name: Arkeo Racing
servers:
  - url: /api
    description: API base path
...
paths:
...
  /races/{raceId}/selection:
    get:
      tags: [Betting]
      summary: Get user's selection
      description: Returns current user's horse selection for the race
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/raceId'
      responses:
        '200':
          description: Current selection
          content:
            application/json:
              schema:
                type: object
                properties:
                  selection:
                    $ref: '#/components/schemas/Selection'
    delete:
      tags: [Betting]
      summary: Clear selection
      description: Removes the current user's horse selection for the race
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/raceId'
      responses:
        '200':
          description: Selection cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  selection:
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Betting]
      summary: Set selection
      description: Select a horse for the race
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/raceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [raceHorseId]
              properties:
                raceHorseId:
                  type: string
      responses:
        '200':
          description: Selection saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Selection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Picks closed
  /races/{raceId}/selections:
    get:
      tags: [Betting]
      summary: Get all selections
      description: Returns all user selections for a race
      parameters:
        - $ref: '#/components/parameters/raceId'
      responses:
        '200':
          description: All selections
          content:
            application/json:
              schema:
                type: object
                properties:
                  selections:
                    type: array
                    items:
                      $ref: '#/components/schemas/SelectionPublic'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: racing_session
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  parameters:
    raceId:
      name: raceId
      in: path
      required: true
      schema:
        type: string
      description: Race UUID
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: invalid_request
...
    Selection:
      type: object
      properties:
        selectionId:
          type: string
        raceHorseId:
          type: string
        createdAt:
          type: string
          format: date-time
    SelectionPublic:
      type: object
      properties:
        selectionId:
          type: string
        raceHorseId:
          type: string
        bettor:
          type: object
          properties:
            nickname:
              type: string
              nullable: true
            walletAddress:
              type: string
              nullable: true
        picks:
          type: array
          items:
            type: object
            properties:
              horseName:
                type: string
...